refactor: improve export feature based on code review
- Extracted file download logic to `utils/download.ts` to reduce component complexity.
- Restored `REFRESH_STRATEGIES` map in `lib/auth.ts` for better extensibility.
- Updated `Content-Type` for FIT exports to `application/vnd.ant.fit`.
- Updated unit tests to reflect these changes.
