+++ lib/auth.ts
@@ -8,6 +8,8 @@ import { env } from './env'
 import { refreshSpotifyToken } from './spotify'
 import { SPOTIFY_DEFAULT_TOKEN_EXPIRY_S } from '@/constants/spotify'
 
+import { SpotifyTokenResponse } from '@/types/core'
+
 // Extend the Session type to include accessToken and error
 declare module 'next-auth' {
   interface Session {
@@ -99,21 +101,27 @@ function getCookieDomain(): string | undefined {
   }
 }
 
+// Define strategies map
+const REFRESH_STRATEGIES: Record<
+  string,
+  (token: string) => Promise<SpotifyTokenResponse>
+> = {
+  spotify: refreshSpotifyToken,
+}
+
 /**
  * Refreshes an expired access token using the refresh token.
- * Supports Spotify provider.
  * Invoked by the NextAuth JWT callback when the access token is expired.
  */
 async function refreshAccessToken(token: JWT) {
-  // Only refresh Spotify tokens for now
-  if (token.provider !== 'spotify') {
+  const refresher = REFRESH_STRATEGIES[token.provider as string]
+
+  if (!refresher) {
     return token
   }
 
   try {
-    const refreshedTokens = await refreshSpotifyToken(
-      token.refreshToken as string
-    )
+    const refreshedTokens = await refresher(token.refreshToken as string)
 
     // Update the token object with new values from the provider
     return {
