+++ tests/unit/app/api/workout/export/[sessionId]/route.test.ts
@@ -110,9 +110,7 @@ describe('API Route: /api/workout/export/[sessionId]', () => {
     const response = await POST(request, context)
 
     expect(response.status).toBe(200)
-    expect(response.headers.get('Content-Type')).toBe(
-      'application/octet-stream'
-    )
+    expect(response.headers.get('Content-Type')).toBe('application/vnd.ant.fit')
     expect(response.headers.get('Content-Disposition')).toContain(
       `attachment; filename="workout_${sessionId}.fit"`
     )
