+++ utils/download.ts
@@ -1,0 +1,16 @@
+/**
+ * Triggers a browser download for a given Blob.
+ *
+ * @param blob - The file data to download.
+ * @param filename - The name of the file to save.
+ */
+export const downloadBlob = (blob: Blob, filename: string): void => {
+  const url = window.URL.createObjectURL(blob)
+  const a = document.createElement('a')
+  a.href = url
+  a.download = filename
+  document.body.appendChild(a)
+  a.click()
+  a.remove()
+  window.URL.revokeObjectURL(url)
+}
